Homepage Stories

+++ database/migrations/2019_02_27_203516_create_users_profiles_table.php
@@ -14,8 +14,8 @@ class CreateUsersProfilesTable extends Migration
     public function up()
     {
         Schema::create('users_profiles', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id')->unique();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id')->unique();
             $table->string('picture', 255)->nullable();
             $table->integer('followers')->default(0);
             $table->integer('videos')->default(0);
