+++ database/migrations/2019_02_28_192940_create_followers_table.php
@@ -14,14 +14,14 @@ class CreateFollowersTable extends Migration
     public function up()
     {
         Schema::create('followers', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('follower_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('follower_id');
 
             $table->foreign('user_id')
             ->references('id')
             ->on('users')
-            ->onDelete('cascade');        
+            ->onDelete('cascade');
 
             $table->foreign('follower_id')
             ->references('id')
